Rename VSUM UI project for uniform naming of UI-related projects
