+++ bundles/framework/tools.vitruv.framework.vsum.ui/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>tools.vitruv.framework.ui.vsum</name>
+	<name>tools.vitruv.framework.vsum.ui</name>
 	<comment></comment>
 	<projects>
 	</projects>
