+++ bundles/framework/tools.vitruv.framework.vsum.ui/META-INF/MANIFEST.MF
@@ -1,10 +1,10 @@
 Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Name: Vitruv Framework UI Virtual Single Underlying Model
-Bundle-SymbolicName: tools.vitruv.framework.ui.vsum;singleton:=true
-Automatic-Module-Name: tools.vitruv.framework.ui.monitorededitor
+Bundle-SymbolicName: tools.vitruv.framework.vsum.ui;singleton:=true
+Automatic-Module-Name: tools.vitruv.framework.vsum.ui
 Bundle-Version: 1.1.0.qualifier
-Bundle-Activator: tools.vitruv.framework.ui.vsum.Activator
+Bundle-Activator: tools.vitruv.framework.vsum.ui.Activator
 Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
  org.eclipse.core.resources;bundle-version="3.10.1",
@@ -20,6 +20,5 @@ Require-Bundle: org.eclipse.ui,
 Bundle-RequiredExecutionEnvironment: JavaSE-11
 Bundle-ActivationPolicy: lazy
 Bundle-Vendor: vitruv.tools
-Export-Package: tools.vitruv.framework.ui.vsum,
- tools.vitruv.framework.ui.vsum.util;x-internal:=true,
- tools.vitruv.framework.ui.vsum.wizard
+Export-Package: tools.vitruv.framework.vsum.ui,
+ tools.vitruv.framework.vsum.ui.wizard
