+++ bundles/framework/tools.vitruv.framework.vsum.ui/build.properties
