+++ bundles/framework/tools.vitruv.framework.vsum.ui/src/tools/vitruv/framework/vsum/ui/Activator.java
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.ui.vsum;
+package tools.vitruv.framework.vsum.ui;
 
 import org.eclipse.ui.plugin.AbstractUIPlugin;
 import org.osgi.framework.BundleContext;
@@ -9,7 +9,7 @@ import org.osgi.framework.BundleContext;
 public class Activator extends AbstractUIPlugin {
 
 	// The plug-in ID
-	public static final String PLUGIN_ID = "VitruvNewWizard"; //$NON-NLS-1$
+	public static final String PLUGIN_ID = "tools.vitruv.framework.vsum.ui"; //$NON-NLS-1$
 
 	// The shared instance
 	private static Activator plugin;
