+++ bundles/framework/tools.vitruv.framework.vsum.ui/src/tools/vitruv/framework/vsum/ui/CreateVsumButtonHandler.xtend
@@ -1,11 +1,11 @@
-package tools.vitruv.framework.ui.vsum
+package tools.vitruv.framework.vsum.ui
 
 import org.eclipse.core.commands.AbstractHandler
 import org.eclipse.core.commands.ExecutionEvent
 import org.eclipse.core.commands.ExecutionException
 import org.eclipse.ui.PlatformUI
 import org.eclipse.jface.wizard.WizardDialog
-import tools.vitruv.framework.ui.vsum.wizard.CreateVsumWizard
+import tools.vitruv.framework.vsum.ui.wizard.CreateVsumWizard
 
 class CreateVsumButtonHandler extends AbstractHandler {
