+++ bundles/framework/tools.vitruv.framework.vsum.ui/src/tools/vitruv/framework/vsum/ui/util/VitruvInstanceCreator.xtend
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.ui.vsum.util
+package tools.vitruv.framework.vsum.ui.util
 
 import tools.vitruv.framework.tuid.TuidManager
 import tools.vitruv.framework.vsum.InternalVirtualModel
