+++ bundles/framework/tools.vitruv.framework.vsum.ui/src/tools/vitruv/framework/vsum/ui/wizard/CreateVsumWizard.java
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.ui.vsum.wizard;
+package tools.vitruv.framework.vsum.ui.wizard;
 
 import org.apache.log4j.Logger;
 import org.eclipse.core.resources.IProject;
@@ -10,7 +10,7 @@ import org.eclipse.ui.IWorkbench;
 
 import tools.vitruv.framework.applications.VitruvApplication;
 import tools.vitruv.framework.domains.VitruvDomain;
-import tools.vitruv.framework.ui.vsum.util.VitruvInstanceCreator;
+import tools.vitruv.framework.vsum.ui.util.VitruvInstanceCreator;
 
 import java.util.Map;
 import java.util.Set;
