+++ bundles/framework/tools.vitruv.framework.vsum.ui/src/tools/vitruv/framework/vsum/ui/wizard/ProjectNamePage.java
@@ -1,4 +1,4 @@
-package tools.vitruv.framework.ui.vsum.wizard;
+package tools.vitruv.framework.vsum.ui.wizard;
 
 import org.eclipse.jface.wizard.WizardPage;
 import org.eclipse.swt.SWT;
